Fix: Regrade all current week battleground contest to use Half PPR
Squashed commit of the following:

commit 260663a11543949f3af5d651ae04e3d555f4a32c
Author: glenn_ferns <glennferns93@gmail.com>
Date:   Tue Nov 2 18:13:09 2021 +0530

    updated date filter

commit abb23b05efa6a9f20a4672a6a4fe794cc0cfc3ea
Author: glenn_ferns <glennferns93@gmail.com>
Date:   Tue Nov 2 18:05:18 2021 +0530

    disabled emails

commit 2e3c38880206b9896f0b36ee748ae45f8fa9db27
Author: glenn_ferns <glennferns93@gmail.com>
Date:   Tue Nov 2 18:01:27 2021 +0530

    removed console.log

commit 3bfd7e9e4a7f3472285a79dff40ce83dbc79d747
Author: glenn_ferns <glennferns93@gmail.com>
Date:   Tue Nov 2 17:59:24 2021 +0530

    changelog
    -  cron to reset the contest data
    - cron to delete gains

+++ src/services/miscellaneous.service.ts
@@ -35,8 +35,52 @@ export class MiscellaneousService {
             // const gains = await this.gainRepository.find({
             //     where: {
             //         contestId: { inq: contestIds },
-                    
+
             //     },
+            // });
+
+            await this.gainRepository.deleteAll({
+                contestId: { inq: contestIds },
+                contestType: 'battleground',
+            });
+
+            const constestData = {
+                winnerId: null,
+                topPropProfit: 0,
+                status: CONTEST_STATUSES.MATCHED,
+                ended: false,
+                endedAt: null,
+                winnerLabel: null,
+                creatorWinAmount: null,
+                claimerWinAmount: null,
+                creatorPlayerFantasyPoints: 0,
+                claimerPlayerFantasyPoints: 0,
+            };
+
+            // @ts-ignore
+            await this.contestRepository.updateAll(constestData, { id: { inq: contestIds } });
+        }
+    }
+
+    async resetNoPPRGradedContests() {
+        /* 
+        Date: 1-10-2021
+        Description: This was run to reset contests that were 
+        graded according to noPPR logic instead of halfPPR */
+
+        const sql =
+            "select id from contest where ended=true and claimerid is not null and createdat > '2021-10-27 00:00:00'";
+
+        const contests = await this.contestRepository.execute(sql, null, null);
+        
+
+        const contestIds = contests.map((contest: { id: number }) => contest.id);
+
+        if (contestIds.length > 0) {
+            logger.info(chalk.redBright(`Contests that were graded incorrectly are `, contestIds));
+
+            // const gains = await this.gainRepository.find({
+            //     where: { contestId: { inq: contestIds }, contestType: 'battleground' },
             // });
 
             await this.gainRepository.deleteAll({
